+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.6.1] - [Unreleased]
+
+### Fixed
+
+#### Checkout
+
+- remove unnecessary useMemo import from useCheckout hook
+
 ## [0.6.0] - 2025-12-08
 
 ### Changed